included maze generation in dungeon generation
