+++ src/engine/DemoDungeon.h
@@ -17,8 +17,6 @@ public:
 
 	virtual void generate(); //sets up layout, reads in scenes, generates rooms
 
-	int **layout; //2d int array representing dungeon structure where -1 is empty and each other int is index into scenes for that room's corrosponding scene
-	vector<string> scenes; //list of paths to scenes representing rooms in dungeon
 private:
 
 };
